Serve the old docs from /docs to keep links alive (#11)
This points at the `embedded-docs` branch on the micrometer-docs repo, so we can test this approach and stage it for deployment. That branch is the build output of the old site, modified to be served from `/docs` without the home page.

Also uses the same workaround we use on micrometer-docs today for redirecting to the client-side routing of the embedded docs site when accessing pages under the `/docs` path.